Compare new diet with current diet

Compare new and current diet in terms of absolute and relative (percent) change

Usage

compare_new_diet(data_new_diet, data_current_diet)

Arguments

data_new_diet

New diet intake

data_current_diet

Current diet that has food name and average intake

Value

A dataframe with food names, current and new diet intake and comparsion metrics