New diet discovery

Find new diet combination that satisfies a set of nutrition and environmental impact constraints.

Usage

find_new_diet(
  diet0,
  diet0_upr,
  diet0_lwr,
  tag_outcomes,
  constraint_val,
  print_runtime = T
)

Arguments

diet0

A vector of current diet intake

diet0_upr

A vector of values that set the upper bound of diet search

diet0_lwr

A vector of values that set the lower bound of diet search

tag_outcomes

A vector of tag_outcome of interest

constraint_val

A list of constraint values that match tag_outcomes

print_runtime

Whether to print runtime. Default is True.

Value

A list of results